Add step to log masked API key in nuget-publish.yml
A new step named "Log masked API key" was added to the
`nuget-publish.yml` file. This step runs if the previous
steps are successful and logs a masked version of the API
key to the console. The API key is partially masked to show
only the first four characters followed by asterisks for
security purposes. The API key is retrieved from the
`GEFCOREHELPER_APIKEY` secret.